python/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.3
aiohttp-retry==2.9.1
aioipfs==0.7.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
base58==2.1.1
bases==0.3.0
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==2.0.12
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
dag-cbor==0.3.3
dnspython==2.7.0
email_validator==2.2.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastcore==1.7.20
frozenlist==1.5.0
fsspec==2024.10.0
gitignore_parser==0.1.9
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
imageio==2.36.0
ipfs-api==0.2.3
ipfshttpclient==0.7.0
ipfspy==0.0.15
ipfsspec==0.5.3
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
morphys==1.0
multiaddr==0.0.9
multidict==6.1.0
multiformats==0.3.1.post4
multiformats-config==0.3.1
netaddr==1.3.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pillow==11.0.0
propcache==0.2.0
pure-protobuf==2.3.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multiformats-cid==0.4.4
py-multihash==2.0.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.27.1
rich==13.9.4
rlp==4.0.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
toolz==1.0.0
typer==0.13.1
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-validation==1.2.11.post4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
varint==1.0.2
watchfiles==0.24.0
web3==7.5.0
websockets==14.1
yarl==1.17.2
